Revert "Dependency: update Clang18 for kernel Build"

+++ kernel/events/core.c
@@ -6859,7 +6859,7 @@ static void perf_fill_ns_link_info(struct perf_ns_link_info *ns_link_info,
 	struct inode *ns_inode;
 	int error;
 
-	error = (int)(long)ns_get_path(&ns_path, task, ns_ops);
+	error = ns_get_path(&ns_path, task, ns_ops);
 	if (!error) {
 		ns_inode = ns_path.dentry->d_inode;
 		ns_link_info->dev = new_encode_dev(ns_inode->i_sb->s_dev);
